+++ experiment.h
@@ -4,6 +4,6 @@
 #include <SDL2/SDL_pixels.h>
 #include "demo.h"
 
-double compare_canvasses(Canvas *A, Canvas *B);
+float compare_canvasses(Canvas *A, Canvas *B);
 
 #endif //ANTI_ALIASING_EXPERIMENT_H
